start merging things
